# backend/docs/swagger/swagger.yaml
basePath: /
...
host: localhost:8180
info:
  contact:
    email: bamort.support@trokan.de
    name: BaMoRT Support
  description: |-
    BaMoRT (MOAM Replacement) - Role-playing Game Character Management System API

    This API provides comprehensive character management, import/export functionality,
    master data management, and PDF generation for tabletop role-playing games.

    ## Authentication
    Most endpoints require authentication via JWT token. Obtain a token by calling POST /login.
    Include the token in the Authorization header: `Authorization: Bearer <token>`

    ## Character Import/Export
    The API supports pluggable adapters for importing/exporting characters from various systems.
    Supported formats include MOAM VTT, Foundry VTT, and more via microservice adapters.
  license:
    name: Custom License
    url: https://github.com/Bardioc26/bamort/blob/main/LICENSE
  title: BaMoRT API
  version: "2.0"
paths:
...
  /api/importer/export/csv/{id}:
    get:
      description: |-
        Exports a character to CSV format (MOAM-compatible)
        Exports a character as a CSV file
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/csv
      - text/csv
      responses:
        "200":
          description: CSV file with character data
          schema:
            type: file
        "400":
          description: Bad request - invalid character ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Character not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error - export failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Export character to CSV
      tags:
      - importer
      - Import/Export
  /api/importer/export/spells/csv:
    get:
      description: |-
        Exports spell master data to CSV format
        Exports all spells as a CSV file
      parameters:
      - description: Game system filter (e.g., 'midgard')
        in: query
        name: game_system
        type: string
      produces:
      - text/csv
      - text/csv
      responses:
        "200":
          description: CSV file with spells
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error - export failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Export spells to CSV
      tags:
      - importer
      - Import/Export
  /api/importer/export/vtt/{id}:
    get:
      description: |-
        Exports a character to VTT JSON format for use in other systems
        Exports a character in Virtual Tabletop (VTT) JSON format
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: Character in VTT format
          schema:
            type: object
        "400":
          description: Bad request - invalid character ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Character not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error - export failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Export character to VTT JSON
      tags:
      - importer
      - Import/Export
  /api/importer/export/vtt/{id}/file:
    get:
      description: |-
        Exports a character to VTT JSON file and returns it as a download
        Downloads a character as a VTT JSON file attachment
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: VTT JSON file download
          schema:
            type: file
        "400":
          description: Bad request - invalid character ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Character not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error - export failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Download character as VTT JSON file
      tags:
      - importer
      - Import/Export
  /api/importer/spells/csv:
    post:
      consumes:
      - multipart/form-data
      - multipart/form-data
      description: |-
        Imports spell data from a CSV file into the database. The CSV file should contain spell information with headers matching the database fields.
        Imports spells from a CSV file
      parameters:
      - description: CSV file to import
        in: formData
        name: file
        required: true
        type: file
      - description: CSV file containing spells
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: Import result
          schema:
            type: object
        "400":
          description: Invalid CSV file
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error - import failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Import spells from CSV
      tags:
      - importer
      - Import/Export
  /api/importer/upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads files for import processing
      parameters:
      - description: File(s) to upload
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Upload result
          schema:
            type: object
        "400":
          description: Invalid file upload
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload files
      tags:
      - Import/Export
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"